Move v1 Hive tests
